% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_data.R
\name{emoji_data}
\alias{emoji_data}
\title{Github release information}
\usage{
emoji_data(version = NULL, path = NULL)
}
\arguments{
\item{path}{Optional path in which to store the data.  If omitted
we use \code{rappdirs} to generate a reasonable path.}
}
\description{
Information to describe how to process github releases
}
