% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_data.R
\name{emoji_data_info}
\alias{emoji_data_info}
\title{Load the v5.0 emoji}
\usage{
emoji_data_info(path = NULL)
}
\arguments{
\item{path}{Optional path in which to store the data.  If omitted
we use \code{rappdirs} to generate a reasonable path.}

\item{version}{Version to load.}
}
\description{
Load v5.0 emoji data
}
\examples{
\dontrun{
baad <- baad_data()
head(baad$data)
}
}
