% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loec.R
\docType{data}
\name{loec}
\alias{loec}
\title{List of Emoji categories}
\format{A list of plausible topics and its emoji keywords:}
\usage{
loec
}
\description{
loec stands for "List of Emoji Categories".
Not an official way to categorize emoji.
Arbitrary categorization done by the author for his own research.
}
\keyword{datasets}
