% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loec.R
\docType{package}
\name{loec}
\alias{loec}
\alias{loec-package}
\title{List of Emoji categories}
\description{
loec stands for "List of Emoji Categories".
Not an official way to categorize emoji.
Arbitrary categorization done by the author for his own research.
}
\keyword{categories}
\keyword{emoji}
